Clean up generation
